module refacor, explicate channel/video search
